+++ src/game/units/wizard/archetypes/psychopath/mod.rs
@@ -1,0 +1,6 @@
+pub(crate) mod constants;
+mod plugin;
+pub(crate) mod run_conditions;
+mod systems;
+
+pub use plugin::PsychopathPlugin;